task4 hw fix
